Add OpenID Connect support; refactor logins
Drop native certificate support on a per-request basis, instead
explicitly log the user in and set a secure cookie which persists for
the session.  Add OpenID Connect support allow users to log in via that.

Add a login.html template for the login page (which currently only
supports OpenID Connect but might support other things later)